moved constants out of config folder, removed config folder
